requirements.txt
absl-py==2.4.0
altair==5.5.0
astunparse==1.6.3
attrs==25.4.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.18.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2025.5.1
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.78.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.13.2
kiwisolver==1.5.0
libclang==18.1.1
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==2.9.0
networkx==3.5
numba==0.64.0
numpy==2.3.1
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
pandas==2.3.3
pillow==11.2.1
protobuf==6.33.0
psycopg2-binary==2.9.10
pyarrow==22.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.37.0
regex==2025.9.1
reportlab==4.4.10
requests==2.32.4
rich==14.3.3
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.16.1
setuptools==80.9.0
shap==0.51.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
SQLAlchemy==2.0.42
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.46.3
wrapt==2.1.2
